+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## v0.3.2
 
+* ✨`[NEW]` Add SOQL scalar function `convertTimezone()`.
 * 📈 Memoize the `immediate-scalar` function calls.
+* 📈 Improves the performance of the standard resolvers.
 * Update README.